fileupload/requirements.txt
django==3.0.7
djangorestframework
djangorestframework-datatables
requests
fuzzywuzzy
